lib: rename fuzzy to jarowinkler
